feat: Add configurable headers
The JWT API might be overlaid with an API key, for example, to prevent
clients from abusing the API, or a malicious third party trying to
perform a DOS attack.
The API key is usually passed through the request headers and it makes
sense to have it available for users to pass.

Signed-off-by: Adrian Oprea <adrian@oprea.rocks>